+++ src/system/application/Controller/identity_controller.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Controller;
+
+class identity_controller
+{
+    use \svc\identity\generate_random_keypair;
+    use \svc\identity\set_identity_info;
+    use \svc\identity\get_identity_info;
+    use \svc\identity\delete_identity_info;
+}