+++ src/system/application/Controller/server_controller.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Controller;
+
+class server_controller
+{
+    use \svc\server\ping;
+}