+++ src/system/boot/services.php
@@ -1,0 +1,5 @@
+<?php
+
+foreach(rsearch(INPHINIT_PATH . "application/Service", ["php"]) as $file) {
+    require_once $file;
+}