+++ src/system/boot/start.php
@@ -29,7 +29,7 @@ foreach(glob(INPHINIT_PATH . "util/*.php") as $file){
 }
 
 require_once INPHINIT_PATH . 'lib/__driver.php';
-
 require_once INPHINIT_PATH . 'boot/errorhandler.php';
+require_once INPHINIT_PATH . 'boot/services.php';
 
 App::exec();