+++ src/system/main.php
@@ -1,13 +1,18 @@
 <?php
 use Inphinit\Routing\Route;
 
-Route::set('GET', '/ping', "ServerController:Ping");
-Route::set('GET', '/generate-keypair', "PrivateKeyController:GetRandomPrivateKey");
+Route::set('GET', '/ping', "server_controller:ping");
 
-Route::set('GET', '/messages', "MessagesController:Browse");
-Route::set('GET', '/messages/receiver/{:[a-zA-Z0-9]+:}{:/?:}', "MessagesController:ReadFromReceiver");
-Route::set('GET', '/messages/sender/{:[a-zA-Z0-9]+:}{:/?:}', "MessagesController:ReadFromSender");
-Route::set('PUT', '/messages/{:[a-zA-Z0-9]+:}{:/?:}', "MessagesController:Edit");
-Route::set('DELETE', '/messages/receiver/{:[a-zA-Z0-9]+:}{:/?:}', "MessagesController:DeleteFromReceiver");
-Route::set('DELETE', '/messages/sender/{:[a-zA-Z0-9]+:}{:/?:}', "MessagesController:DeleteFromSender");
-Route::set('POST', '/messages', "MessagesController:Add");
+Route::set('GET', '/identity/generate-keypair', "identity_controller:generate_random_keypair");
+Route::set('POST', '/identity{:/?:}', "identity_controller:set_identity_info");
+Route::set('GET', '/identity{:/?:}', "identity_controller:get_identity_info");
+Route::set('DELETE', '/identity{:/?:}', "identity_controller:delete_identity_info");
+
+Route::set('GET', '/messages{:/?:}', "messages_controller:browse");
+Route::set('GET', '/messages/receiver/{:[a-zA-Z0-9]+:}{:/?:}', "messages_controller:read_from_receiver");
+Route::set('GET', '/messages/sender/{:[a-zA-Z0-9]+:}{:/?:}', "messages_controller:read_from_sender");
+Route::set('PUT', '/messages/{:[a-zA-Z0-9]+:}{:/?:}', "messages_controller:edit");
+Route::set('DELETE', '/messages/receiver/{:[a-zA-Z0-9]+:}{:/?:}', "messages_controller:delete_from_receiver");
+Route::set('DELETE', '/messages/sender/{:[a-zA-Z0-9]+:}{:/?:}', "messages_controller:delete_from_sender");
+Route::set('POST', '/messages{:/?:}', "messages_controller:add");
+